+++ src/views/WebProfileBuilder/Interactive.vue
@@ -1,0 +1,11 @@
+<script setup lang="ts">
+
+</script>
+
+<template>
+
+</template>
+
+<style>
+
+</style>
